+++ packages/components/src/index.tsx
@@ -1,5 +1,6 @@
 import { ReactElement } from 'react';
 import { render } from 'react-dom';
+import $ from 'jquery';
 
 export * from './container.component';
 export * from './dialog.component';
@@ -18,3 +19,11 @@ export function launch(viewport: ReactElement) {
   document.body.appendChild(root);
   render(viewport, root);
 }
+
+$(function() {
+  // $(document).on('click', '.dropdown-toggle', function(e) { console.log('.dropdown-toggle'); $(e.target).parent('.dropdown').children('.dropdown-menu').slideToggle() });
+  // $(document).on('click', e => { console.log('click outside'); $('.dropdown-menu').slideToggle() });
+  // $(document).on('click', '.dropdown-item', e => { console.log('click inside'); $('.dropdown-menu').slideUp() });
+  // $(document).on('click', '.dropdown-multi-select .dropdown-item', e => { console.log('.dropdown-multi-select'); e.stopPropagation(); });
+  // $(document).on('click.bs.dropdown.data-api', '.dropdown', function(e) { console.log($(e.target).hasClass('dropdown-multi-select')); return !$(e.target).hasClass('dropdown-multi-select') });
+});